requirements.txt
brax==0.10.5
chex==0.1.86
flax==0.9.0
hydra-core==1.3.2
jax==0.4.33
jaxlib==0.4.33
matplotlib==3.7.5
numpy==1.26.4
optax==0.2.3
pandas==1.5.2
PyYAML==6.0
qdax==0.4.0
seaborn==0.12.2
scipy==1.12.0
toolz==0.12.0
tqdm==4.66.5
typing_extensions==4.4.0
